Regroup tests to align with source code
